Extract `createValidatorBlock`
